Diagrams.Core.Names

Description

This module defines a type of names which can be used for referring to subdiagrams, and related types.

Names

Atomic names

data AName where

Atomic names. AName is just an existential wrapper around things which are Typeable, Ord and Show.

Constructors

AName :: (Typeable a, Ord a, Show a) => a -> AName

Names

newtype Name

A (qualified) name is a (possibly empty) sequence of atomic names.

Constructors

Name [AName]

class (Typeable a, Ord a, Show a) => IsName a where

Class for those types which can be used as names. They must support Typeable (to facilitate extracting them from existential wrappers), Ord (for comparison and efficient storage) and Show.

Methods

toName :: a -> Name

(.>) :: (IsName a1, IsName a2) => a1 -> a2 -> Name

Convenient operator for writing qualified names with atomic components of different types. Instead of writing toName a1 <> toName a2 <> toName a3 you can just write a1 .> a2 .> a3.

Qualifiable

class Qualifiable q where

Instances of Qualifiable are things which can be qualified by prefixing them with a name.

Methods

(|>) :: IsName a => a -> q -> q

Qualify with the given name.