Diagrams.Core.Points

Description

A type for points (as distinct from vectors).

Points

newtype Point f a :: (* -> *) -> * -> *

A handy wrapper to help distinguish points from vectors at the type level

Constructors

P (f a)

origin :: (Additive f, Num a) => Point f a

Vector spaces have origins.

(*.) :: (Functor v, Num n) => n -> Point v n -> Point v n

Scale a point by a scalar. Specialized version of '(*^)'.

relative :: (Additive f, Num a) => Point f a -> Iso' (Point f a) (f a)

An isomorphism between points and vectors, given a reference point.

_Point :: (Profunctor p, Functor f) => p (f a) (f (f a)) -> p (Point f a) (f (Point f a))

reflectThrough :: (Additive v, Num n) => Point v n -> Point v n -> Point v n

Mirror a point through a given point.

mirror :: (Additive v, Num n) => Point v n -> Point v n

relative2 :: (Additive v, Num n) => Point v n -> (v n -> v n -> v n) -> Point v n -> Point v n -> Point v n

Apply a transformation relative to the given point.

relative3 :: (Additive v, Num n) => Point v n -> (v n -> v n -> v n -> v n) -> Point v n -> Point v n -> Point v n -> Point v n

Apply a transformation relative to the given point.