Diagrams.Core.Trace

Description

diagrams-core defines the core library of primitives forming the basis of an embedded domain-specific language for describing and rendering diagrams.

The Trace module defines a data type and type class for "traces", aka functional boundaries, essentially corresponding to embedding a raytracer with each diagram.

Traces

newtype Trace v

Every diagram comes equipped with a trace. Intuitively, the trace for a diagram is like a raytracer: given a line (represented as a base point and a direction), the trace computes the distance from the base point along the line to the first intersection with the diagram. The distance can be negative if the intersection is in the opposite direction from the base point, or infinite if the ray never intersects the diagram. Note: to obtain the distance to the furthest intersection instead of the closest, just negate the direction vector and then negate the result.

Note that the output should actually be interpreted not as an absolute distance, but as a multiplier relative to the input vector. That is, if the input vector is v and the returned scalar is s, the distance from the base point to the intersection is given by s * magnitude v.

Constructors

Trace

Fields

appTrace :: Point v -> v -> PosInf (Scalar v)

inTrace :: ((Point v -> v -> PosInf (Scalar v)) -> Point v -> v -> PosInf (Scalar v)) -> Trace v -> Trace v

mkTrace :: (Point v -> v -> PosInf (Scalar v)) -> Trace v

Traced class

class (Ord (Scalar (V a)), VectorSpace (V a)) => Traced a where

Traced abstracts over things which have a trace.

Methods

getTrace :: a -> Trace (V a)

Compute the trace of an object.

Computing with traces

traceV :: Traced a => Point (V a) -> V a -> a -> Maybe (V a)

Compute the vector from the given point to the boundary of the given object in the given direction, or Nothing if there is no intersection.

traceP :: Traced a => Point (V a) -> V a -> a -> Maybe (Point (V a))

Given a base point and direction, compute the closest point on the boundary of the given object, or Nothing if there is no intersection in the given direction.

maxTraceV :: Traced a => Point (V a) -> V a -> a -> Maybe (V a)

Like traceV, but computes a vector to the *furthest* point on the boundary instead of the closest.

maxTraceP :: Traced a => Point (V a) -> V a -> a -> Maybe (Point (V a))

Like traceP, but computes the *furthest* point on the boundary instead of the closest.