Graphics.Rendering.Diagrams.MList

Description

Heterogeneous lists of monoids.

Heterogeneous monoidal lists

The idea of heterogeneous lists has been around for a long time. Here, we adopt heterogeneous lists where the element types are all monoids: this allows us to leave out identity values, so that a heterogeneous list containing only a single non-identity value can be created without incurring constraints due to all the other types, by leaving all the other values out.

data Nil

The empty heterogeneous list.

Constructors

Nil

data a ::: l

Cons for heterogeneous lists.

Constructors

Missing l
The a value is missing, and should be construed as mempty.

a ::: l
An a value followed by a heterogeneous list l.

class MList l where

Type class for heterogeneous monoidal lists, with a single method allowing construction of an empty list.

Methods

empty :: l

The empty heterogeneous list of type l. Of course, empty == mempty, but unlike mempty, empty does not require Monoid constraints on all the elements of l.

Converting to tuples

type family Tuple l :: *

A type function to compute the tuple-based representation for instances of MList.

class ToTuple l where

toTuple can be used to convert a heterogeneous list to its tuple-based representation.

Methods

toTuple :: l -> Tuple l

Accessing embedded values

class l :>: a where

The relation l :>: a holds when a is the type of an element in l. For example, (Char ::: Int ::: Bool ::: Nil) :>: Int.

Methods

inj :: a -> l

Inject a value into an otherwise empty heterogeneous list.

get :: l -> a

Get the value of type a from a heterogeneous list.

alt :: (a -> a) -> l -> l

Alter the value of type a by applying the given function to it.

Monoid actions of heterogeneous lists

Monoidal heterogeneous lists may act on one another as you would expect, with each element in the first list acting on each in the second. Unfortunately, coding this up in type class instances is a bit fiddly.

newtype SM m

SM, an abbreviation for "single monoid" (as opposed to a heterogeneous list of monoids), is only used internally to help guide instance selection when defining the action of heterogeneous monoidal lists on each other.

Constructors

SM m