Diagrams.Core.Measure

Documentation

newtype Measured n a

'Measured n a' is an object that depends on local, normalized and global scales. The normalized and global scales are calculated when rendering a diagram.

For attributes, the local scale gets multiplied by the average scale of the transform.

Constructors

Measured

Fields

unmeasure :: (n, n, n) -> a

type Measure n = Measured n n

A measure is a Measured number.

fromMeasured :: Num n => n -> n -> Measured n a -> a

fromMeasured globalScale normalizedScale measure -> a

output :: n -> Measure n

Output units don't change.

local :: Num n => n -> Measure n

Local units are scaled by the average scale of a transform.

global :: Num n => n -> Measure n

Global units are scaled so that they are interpreted relative to the size of the final rendered diagram.

normalized :: Num n => n -> Measure n

Normalized units get scaled so that one normalized unit is the size of the final diagram.

normalised :: Num n => n -> Measure n

Just like normalized but spelt properly.

scaleLocal :: Num n => n -> Measured n a -> Measured n a

Scale the local units of a Measured thing.

atLeast :: Ord n => Measure n -> Measure n -> Measure n

Calculate the smaller of two measures.

atMost :: Ord n => Measure n -> Measure n -> Measure n

Calculate the larger of two measures.