Diagrams.Core.Query

Description

The Query module defines a type for "queries" on diagrams, which are functions from points in a vector space to some monoid.

Documentation

newtype Query v m

A query is a function that maps points in a vector space to values in some monoid. Queries naturally form a monoid, with two queries being combined pointwise.

The idea for annotating diagrams with monoidal queries came from the graphics-drawingcombinators package, http://hackage.haskell.org/package/graphics-drawingcombinators.

Constructors

Query

Fields

runQuery :: Point v -> m