Diagrams.Core.Transform

Description

Diagrams defines the core library of primitives forming the basis of an embedded domain-specific language for describing and rendering diagrams.

The Transform module defines generic transformations parameterized by any vector space.

Transformations

Invertible linear transformations

data u :-: v infixr 7

(v1 :-: v2) is a linear map paired with its inverse.

Constructors

(u -> v) :-: (v -> u) infixr 7

(<->) :: (u -> v) -> (v -> u) -> u :-: v

Create an invertible linear map from two functions which are assumed to be linear inverses.

linv :: (u :-: v) -> v :-: u

Invert a linear map.

lapp :: (u :-: v) -> u -> v

Apply a linear map to a vector.

General transformations

data Transformation v n

General (affine) transformations, represented by an invertible linear map, its transpose, and a vector representing a translation component.

By the transpose of a linear map we mean simply the linear map corresponding to the transpose of the map's matrix representation. For example, any scale is its own transpose, since scales are represented by matrices with zeros everywhere except the diagonal. The transpose of a rotation is the same as its inverse.

The reason we need to keep track of transposes is because it turns out that when transforming a shape according to some linear map L, the shape's normal vectors transform according to L's inverse transpose. (For a more detailed explanation and proof, see https://wiki.haskell.org/Diagrams/Dev/Transformations.) This is exactly what we need when transforming bounding functions, which are defined in terms of perpendicular (i.e. normal) hyperplanes.

For more general, non-invertible transformations, see Diagrams.Deform (in diagrams-lib).

Constructors

Transformation (v n :-: v n) (v n :-: v n) (v n)

inv :: (Functor v, Num n) => Transformation v n -> Transformation v n

Invert a transformation.

transp :: Transformation v n -> v n :-: v n

Get the transpose of a transformation (ignoring the translation component).

transl :: Transformation v n -> v n

Get the translational component of a transformation.

dropTransl :: (Additive v, Num n) => Transformation v n -> Transformation v n

Drop the translational component of a transformation, leaving only the linear part.

apply :: Transformation v n -> v n -> v n

Apply a transformation to a vector. Note that any translational component of the transformation will not affect the vector, since vectors are invariant under translation.

papply :: (Additive v, Num n) => Transformation v n -> Point v n -> Point v n

Apply a transformation to a point.

fromLinear :: (Additive v, Num n) => (v n :-: v n) -> (v n :-: v n) -> Transformation v n

Create a general affine transformation from an invertible linear transformation and its transpose. The translational component is assumed to be zero.

fromOrthogonal :: (Additive v, Num n) => (v n :-: v n) -> Transformation v n

An orthogonal linear map is one whose inverse is also its transpose.

fromSymmetric :: (Additive v, Num n) => (v n :-: v n) -> Transformation v n

A symmetric linear map is one whose transpose is equal to its self.

basis :: (Additive t, Traversable t, Num a) => [t a]

Produce a default basis for a vector space. If the dimensionality of the vector space is not statically known, see basisFor.

dimension :: forall a. (Additive (V a), Traversable (V a)) => a -> Int

Get the dimension of an object whose vector space is an instance of HasLinearMap, e.g. transformations, paths, diagrams, etc.

onBasis :: (Additive v, Traversable v, Num n) => Transformation v n -> ([v n], v n)

Get the matrix equivalent of the linear transform, (as a list of columns) and the translation vector. This is mostly useful for implementing backends.

listRep :: Foldable v => v n -> [n]

Convert a vector v to a list of scalars.

matrixRep :: (Additive v, Traversable v, Num n) => Transformation v n -> [[n]]

Convert the linear part of a Transformation to a matrix representation as a list of column vectors which are also lists.

matrixHomRep :: (Additive v, Traversable v, Num n) => Transformation v n -> [[n]]

Convert a `Transformation v` to a homogeneous matrix representation. The final list is the translation. The representation leaves off the last row of the matrix as it is always [0,0, ... 1] and this representation is the defacto standard for backends.

determinant :: (Additive v, Traversable v, Num n) => Transformation v n -> n

The determinant of (the linear part of) a Transformation.

isReflection :: (Additive v, Traversable v, Num n, Ord n) => Transformation v n -> Bool

Determine whether a Transformation includes a reflection component, that is, whether it reverses orientation.

avgScale :: (Additive v, Traversable v, Floating n) => Transformation v n -> n

Compute the "average" amount of scaling performed by a transformation. Satisfies the properties

  avgScale (scaling k) == k
  avgScale (t1 <> t2)  == avgScale t1 * avgScale t2
  

eye :: (HasBasis v, Num n) => v (v n)

Identity matrix.

The Transformable class

type HasLinearMap v = (HasBasis v, Traversable v)

HasLinearMap is a constraint synonym, just to help shorten some of the ridiculously long constraint sets.

type HasBasis v = (Additive v, Representable v, Rep v ~ E v)

An Additive vector space whose representation is made up of basis elements.

class Transformable t where

Type class for things t which can be transformed.

Methods

transform :: Transformation (V t) (N t) -> t -> t

Apply a transformation to an object.

Translational invariance

newtype TransInv t

TransInv is a wrapper which makes a transformable type translationally invariant; the translational component of transformations will no longer affect things wrapped in TransInv.

Constructors

TransInv t

Vector space independent transformations

Most transformations are specific to a particular vector space, but a few can be defined generically over any vector space.

translation :: v n -> Transformation v n

Create a translation.

translate :: Transformable t => Vn t -> t -> t

Translate by a vector.

scaling :: (Additive v, Fractional n) => n -> Transformation v n

Create a uniform scaling transformation.

scale :: (InSpace v n a, Eq n, Fractional n, Transformable a) => n -> a -> a

Scale uniformly in every dimension by the given scalar.