Diagrams.Core.V

Description

Type family for identifying associated vector spaces.

Documentation

type family V a :: *

Many sorts of objects have an associated vector space in which they "live". The type function V maps from object types to the associated vector space.