Diagrams.Core.V

Description

Type family for identifying associated vector spaces.

Documentation

type family V a :: * -> *

Many sorts of objects have an associated vector space in which they "live". The type function V maps from object types to the associated vector space. The resulting vector space has kind * -> * which means it takes another value (a number) and returns a concrete vector. For example V2 has kind * -> * and V2 Double is a vector.

type family N a :: *

The numerical field for the object, the number type used for calculations.

type Vn a = V a (N a)

Convenient type alias to retrieve the vector type associated with an object's vector space. This is usually used as Vn a ~ v n where v is the vector space and n is the numerical field.

type InSpace v n a = (V a ~ v, N a ~ n, Additive v, Num n)

InSpace v n a means the type a belongs to the vector space v n, where v is Additive and n is a Num.

type SameSpace a b = (V a ~ V b, N a ~ N b)

SameSpace a b means the types a and b belong to the same vector space v n.